Movendo funcões utitilitárias para um módulo próprio
